#18 [FIX] 시간행 크기 고정
